fix: simplify the exported text file's contents
The randbat data cannot be used properly to create teams, as the data is mismatched from what is allowed
in all Pokemon Showdown formats, causing you to have to manually fix the data. So instead, we can just print
a list of Pokemon names and use that to import teams instead, and manually add the data we want for our teams.